CSP Corrections
Further content security policy fixes... found the issue with the SVG images as well, proved it with the wordmark here, will update others tomorrow. Existing errors in console happen with original images as well. (tested with historical working files, then proved it with wordmark updated file) Does not appear to actually stop the image from loading, as it's a style error, not image.